完成GroupAdmin改密码,修改check_request小bug,删除ResumeList的print
