--- client/tpl/group_list.html
@@ -1,64 +1,0 @@
-<head>
-    <style type="text/css"> 
-        .cv-box {
-            width: 100%; 
-        }
-        .cv-content {
-            position: relative;
-            max-height: 80px;
-            overflow: hidden;  
-            text-overflow: ellipsis;  
-            display: -webkit-box;  
-            -webkit-line-clamp: 3;  
-            -webkit-box-orient: vertical;
-        }
-        
-        .group-cvlist-table {
-            margin 0;
-            padding: 0px;
-        }
-        .group-cvlist{
-            padding-left: 5px;
-        }
-    </style>
-</head>
-<div class="container" ng-init="show=false">
-    <h2 class="form-signin-heading">查看简历 <a href="#/index/" class="small">返回首页</a></h2>
-    <div class="row">
-        <div class="col-xs-4">
-            <div class="list-group">
-                <a href="/group/list/" class="list-group-item active">简历列表</a>
-                <a href="/group/admin/" class="list-group-item">管理员列表</a>
-                <a href="" class="list-group-item">授权码</a>
-                <a href="" class="list-group-item">修改密码</a>
-            </div>
-        </div>
-        <div class="col-xs-8">
-            <div ng-show="show" class="list">
-                <div ng-repeat="item in items" >
-                    {{getInfo(item)}}
-                    <div class="cv-box panel panel-default">
-                        <div class="panel-heading">
-                            <td>Email(测试):<span ng-bind="item"></span></td>
-                            <table class="group-cvlist-table">
-                                <tr>
-                                    <td>姓名:<span ng-bind="name"></span></td>
-                                    <td class="group-cvlist">QQ:<span ng-bind="qq"></span></td>
-                                    <td class="group-cvlist"><a href="" class="red">允许</a></td>
-                                    <td class="group-cvlist"><a href="" class="red">拒绝</a></td>
-                                    <td class="group-cvlist"><a href="" class="red">我的评分</a></td>
-                                    <td class="group-cvlist"><a href="" class="red">平均:</a></td>
-                                </tr>
-                            </table>
-                        </div>
-                        <div class="panel-body">
-                            <p class="cv-content"><span ng-bind="content"></span></p>
-                            <div class="text_center"><a href="">更多(点此下拉,查看更多内容)</a></div>
-                        </div>
-                    </div>
-                </div>
-                
-            </div>
-        </div>
-    </div>   
-</div>
