+++ server/api/controller/group/form.py
@@ -1,0 +1,23 @@
+#-*- coding:utf-8 -*-
+
+from django.forms import (Form, CharField, PasswordInput, IntegerField, ChoiceField)
+
+
+
+
+class PwdForm(Form):
+    password = CharField(label=u'密码：', widget=PasswordInput())
+
+class MngResumeForm(Form):
+    STATUS_CHOICES = (
+        (0,u'申请中'),
+        (1,u'允许的'),
+        (2,u'拒绝的'),
+        (3,u'拉黑的'),
+    )
+    resumeId = IntegerField()
+    status = ChoiceField(choices=STATUS_CHOICES, required=False)
+    rank = IntegerField(required=False)
+
+class DelResumeForm(Form):
+    resumeId = IntegerField()
